& stop throwing error messages on native docker hosts
To determine the correct host, we have to check if docker-machine is
running. On Windows hosts, the command `active` returns 0 and then
executes the command `ip` to determine the IP of the virtual machine
that runs the Docker host. On hosts that run Docker natively, the host
will always be localhost.